Moved UniTaskUtils and replaced defines with a bool
