+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/AssetManager/Gif/GifProcessor/GifProcessor.asmdef
@@ -1,26 +1,27 @@
 {
-  "name": "GifProcessor",
-  "rootNamespace": "",
-  "references": [
-    "GUID:b82364d78784241f8afff9cbbf8bae9f",
-    "GUID:4efd4f6072cb64982ab90a7168745f82",
-    "GUID:0a33bdeafa5c84cb1b5534b5feb3c773",
-    "GUID:c44f6fd10d3d94432a107d581e0096b5",
-    "GUID:b1087c5731ff68448a0a9c625bb7e52d",
-    "GUID:e9cee4050cfa5534d9b26e2782df20be",
-    "GUID:e99faafecbf8dd74d8130597511b9b27",
-    "GUID:ee6523961ef177343ad763fcd55c7c46",
-    "GUID:fbcc413e192ef9048811d47ab0aca0c0", 
-    "GUID:f51ebe6a0ceec4240a699833d6309b23",
-    "GUID:3b80b0b562b1cbc489513f09fc1b8f69"
-  ],
-  "includePlatforms": [],
-  "excludePlatforms": [],
-  "allowUnsafeCode": false,
-  "overrideReferences": true,
-  "precompiledReferences": [],
-  "autoReferenced": true,
-  "defineConstraints": [],
-  "versionDefines": [],
-  "noEngineReferences": false
+    "name": "GifProcessor",
+    "rootNamespace": "",
+    "references": [
+        "GUID:b82364d78784241f8afff9cbbf8bae9f",
+        "GUID:4efd4f6072cb64982ab90a7168745f82",
+        "GUID:0a33bdeafa5c84cb1b5534b5feb3c773",
+        "GUID:c44f6fd10d3d94432a107d581e0096b5",
+        "GUID:b1087c5731ff68448a0a9c625bb7e52d",
+        "GUID:e9cee4050cfa5534d9b26e2782df20be",
+        "GUID:e99faafecbf8dd74d8130597511b9b27",
+        "GUID:ee6523961ef177343ad763fcd55c7c46",
+        "GUID:fbcc413e192ef9048811d47ab0aca0c0",
+        "GUID:f51ebe6a0ceec4240a699833d6309b23",
+        "GUID:8aa60ff897c74401cb96cdbc583fed0d",
+        "GUID:3b80b0b562b1cbc489513f09fc1b8f69"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": true,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
 }
