+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Helpers/Utils/UniTask/UniTaskUtils.asmdef
@@ -1,0 +1,18 @@
+{
+    "name": "UniTaskUtils",
+    "rootNamespace": "",
+    "references": [
+        "GUID:28f74c468a54948bfa9f625c5d428f56",
+        "GUID:f51ebe6a0ceec4240a699833d6309b23",
+        "GUID:1e6b57fe78f7b724e9567f29f6a40c2c"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
