+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Helpers/Utils/UniTask/UniTaskUtils.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8aa60ff897c74401cb96cdbc583fed0d
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
