+++ unity-renderer/Assets/UnityGLTF/Scripts/UnityGLTFAssembly.asmdef
@@ -18,6 +18,7 @@
         "GUID:2995626b54c60644988f134a69a77450",
         "GUID:1e6b57fe78f7b724e9567f29f6a40c2c",
         "GUID:f51ebe6a0ceec4240a699833d6309b23",
+        "GUID:8aa60ff897c74401cb96cdbc583fed0d",
         "GUID:3b80b0b562b1cbc489513f09fc1b8f69"
     ],
     "includePlatforms": [],
